+++ vite.config.ts
@@ -26,7 +26,6 @@ export default defineConfig(async () => ({
     rollupOptions: {
       input: {
         main: fileURLToPath(new URL("./index.html", import.meta.url)),
-        settings: fileURLToPath(new URL("./settings.html", import.meta.url)),
       },
     },
   },